feat: add driver license - remove old UpdateDriverLicense for updated database
